Project Beats: move player interfaces to separate files
