+++ apps/src/music/player/SamplePlayer.ts
@@ -1,5 +1,5 @@
+import { Effects } from './interfaces/Effects';
 import MusicLibrary from './MusicLibrary';
-import {Effects} from './MusicPlayer';
 
 // Using require() to import JS in TS files
 const soundApi = require('./sound');
