+++ apps/src/music/player/interfaces/Effects.ts
@@ -1,0 +1,10 @@
+/**
+ * Describes effects that can be applied to playback events.
+ */
+export interface Effects {
+  volume?: EffectValue;
+  filter?: EffectValue;
+  delay?: EffectValue;
+}
+
+type EffectValue = 'normal' | 'medium' | 'low';
