+++ apps/src/music/player/interfaces/PlaybackEvent.ts
@@ -1,0 +1,20 @@
+import {Effects} from './Effects';
+import {FunctionContext} from './FunctionContext';
+import {SkipContext} from './SkipContext';
+
+export interface PlaybackEvent {
+  /** Type of event */
+  type: 'sound' | 'pattern';
+  /** Measure when this event occurs */
+  when: number;
+  /** Whether this event was triggered or scheduled via standard playback */
+  triggered: boolean;
+  /** ID of the track this event belongs to (only used in Tracks mode) */
+  trackId?: string;
+  /** Function context this event belongs to (only used in Simple2 mode) */
+  functionContext?: FunctionContext;
+  /** Skip context for this event (whether it is skippable and is currently being skipped) */
+  skipContext?: SkipContext;
+  /** Effects applied to this event */
+  effects?: Effects;
+}
