+++ apps/src/music/player/interfaces/SkipContext.ts
@@ -1,0 +1,7 @@
+/**
+ * Describes whether a sound is inside a random block, and is currently being skipped.
+ */
+export interface SkipContext {
+  insideRandom: boolean;
+  skipSound: boolean;
+}
