+++ apps/src/music/player/interfaces/SoundEvent.ts
@@ -1,0 +1,9 @@
+import {PlaybackEvent} from './PlaybackEvent';
+
+/**
+ * A playback event representing a single sound
+ */
+export interface SoundEvent extends PlaybackEvent {
+  type: 'sound';
+  id: string;
+}
